ClassWork

+++ src/main/java/dtoSuper/RecordDto.java
@@ -18,9 +18,9 @@ public class RecordDto {
     private String timeFrom;
     private String timeTo;
     private String title;
-    private int totalSalary;
+    private float totalSalary;
     private String type;
-    private int wage;
+    private float wage;
 
 
 }
